+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/tabstray/TabsTrayPresenter.kt
@@ -14,6 +14,7 @@ import kotlinx.coroutines.flow.map
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.browser.thumbnails.ThumbnailsUseCases
 import mozilla.components.concept.tabstray.Tabs
 import mozilla.components.concept.tabstray.TabsTray
 import mozilla.components.feature.tabs.ext.toTabs
@@ -27,6 +28,7 @@ import mozilla.components.support.ktx.kotlinx.coroutines.flow.ifChanged
 class TabsTrayPresenter(
     private val tabsTray: TabsTray,
     private val store: BrowserStore,
+    private val thumbnailsUseCases: ThumbnailsUseCases,
     internal var tabsFilter: (TabSessionState) -> Boolean,
     private val closeTabsTray: () -> Unit
 ) {
@@ -43,6 +45,12 @@ class TabsTrayPresenter(
 
     private suspend fun collect(flow: Flow<BrowserState>) {
         flow.map { it.toTabs(tabsFilter) }
+            .map { tabs ->
+                // Load the tab thumbnail from the memory or disk caches.
+                tabs.copy(list = tabs.list.map { tab ->
+                    tab.copy(thumbnail = thumbnailsUseCases.loadThumbnail(tab.id))
+                })
+            }
             .ifChanged()
             .collect { tabs ->
                 // Do not invoke the callback on start if this is the initial state.